requirements.txt
# AI Fit Check - Dependencies
# ============================

# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0

# Grounded SAM 2 - Clothing Segmentation
# Install separately: pip install segment-anything-2
groundingdino-py>=0.4.0
supervision>=0.21.0

# Real-ESRGAN - Super Resolution
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8

# CharacterGen - Multi-view Generation
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0

# IC-Light - Lighting Harmonization
# Uses diffusers (already included above)

# Utilities
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-image>=0.21.0
httpx>=0.27.0  # async HTTP client for Fashn.ai API
